+++ backend/app/usecase/match/contains_all.go
@@ -1,4 +1,4 @@
-package matcher
+package match
 
 import "strings"
