+++ backend/app/usecase/match/contains_any.go
@@ -1,4 +1,4 @@
-package matcher
+package match
 
 import "strings"
